requirements.txt
torch>=2.0,<3.0
torchvision>=0.15
timm>=0.8
opencv-python>=4.6
albumentations>=1.3
pandas>=1.4
pyyaml>=6.0
tqdm>=4.64
Pillow>=9.0
grad-cam==1.5.5
scikit-learn>=1.1
fastapi>=0.95
uvicorn>=0.22
matplotlib>=3.5
seaborn>=0.12
optuna>=3.0
psutil>=5.9
optuna